random_forest/requirements.txt
numpy>=1.24
scikit-learn>=1.3
huggingface_hub>=0.15
datasets>=2.18
scikit-image>=0.22
huggingface-hub
matplotlib>=3.8
pillow>=10.0
joblib>=1.3
flask>=3.0
python-dotenv
albumentations>=1.3.0
opencv-python>=4.8.0
torch>=2.0.0